+++ clang/docs/ReleaseNotes.rst
@@ -412,6 +412,10 @@ Modified Compiler Flags
   ``memset`` and similar functions for which it is a documented undefined
   behavior.
 
+- Added ``-fmodules-reduced-bmi`` flags corresponding to
+  ``-fexperimental-modules-reduced-bmi`` flag and emit warning for not emitting
+  reduced BMI.
+
 Removed Compiler Flags
 -------------------------